Fix: media from previous posts being cached
Media from previous posts were being cached with the latest post. This
creates a situation where new posts are being embedded with all
previous media. This was because we were downloading media for all
items before we determine if they've been cached. We should only
download media for uncached items.

This was fixed incidentally as part of a refactoring here:

https://github.com/actionslaw/actionslaw-action/commit/3a3c56da1f243b1de7f9be85290f6e9b85b74814